+++ deepchem/feat/molecule_featurizers/dmpnn_featurizer.py
@@ -318,7 +318,7 @@ def generate_global_features(mol: RDKitMol,
     >>> nonzero_features_indices
     array([1264])
     >>> global_features[nonzero_features_indices[0]]
-    1.0
+    np.float64(1.0)
 
     """
     global_features: List[np.ndarray] = []
